+++ src/main/java/blackjack/domain/card/Cards.java
@@ -1,6 +1,7 @@
-package blackjack;
+package blackjack.domain.card;
 
 
+import blackjack.domain.card.Card;
 import java.util.List;
 
 public class Cards {
