+++ src/main/java/blackjack/domain/common/Name.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain.common;
 
 public class Name {
     private String value;
