+++ src/main/java/blackjack/domain/player/CardReceivable.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain.player;
 
 public interface CardReceivable {
     public boolean isReceivable();
