+++ src/main/java/blackjack/domain/player/GamePlayer.java
@@ -1,5 +1,8 @@
-package blackjack;
+package blackjack.domain.player;
 
+
+import blackjack.domain.common.Name;
+import blackjack.domain.card.Cards;
 
 public class GamePlayer extends Player implements CardReceivable {
     private static final Integer RECEIVE_SIZE = 21;
