+++ src/main/java/blackjack/domain/result/DealerResult.java
@@ -1,5 +1,6 @@
-package blackjack;
+package blackjack.domain.result;
 
+import blackjack.domain.common.Name;
 import java.util.EnumMap;
 import java.util.List;
