+++ src/main/java/blackjack/domain/result/GamePlayerResult.java
@@ -1,8 +1,10 @@
-package blackjack;
+package blackjack.domain.result;
+
+import blackjack.domain.common.Name;
 
 public class GamePlayerResult {
-    Name name;
-    ResultStatus resultStatus;
+    private final Name name;
+    private final ResultStatus resultStatus;
 
     public GamePlayerResult(Name name, ResultStatus resultStatus) {
         this.name = name;
