+++ src/test/java/blackjack/GamePlayerTest.java
@@ -4,6 +4,7 @@ package blackjack;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 
+import blackjack.domain.card.CardValue;
 import blackjack.fixture.PlayerFixture;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
