+++ src/test/java/blackjack/NameTest.java
@@ -2,6 +2,7 @@ package blackjack;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import blackjack.domain.common.Name;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
