+++ src/test/java/blackjack/ResultTest.java
@@ -2,6 +2,11 @@ package blackjack;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import blackjack.domain.common.Name;
+import blackjack.domain.result.DealerResult;
+import blackjack.domain.result.GamePlayerResult;
+import blackjack.domain.result.Result;
+import blackjack.domain.result.ResultStatus;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
