+++ src/test/java/blackjack/fixture/CardFixture.java
@@ -1,9 +1,9 @@
 package blackjack.fixture;
 
-import blackjack.Card;
-import blackjack.CardSymbol;
-import blackjack.CardValue;
-import blackjack.Cards;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardSymbol;
+import blackjack.domain.card.CardValue;
+import blackjack.domain.card.Cards;
 import java.util.List;
 
 public class CardFixture {
